### 2024-02-04 #6319 @editorialbot
[PRE REVIEW]: PyBWE: Python tools for Bandwidth Extrapolation of planetary radar signals
**Submitting author:** @NicOudart (<a href="http://orcid.org/0000-0002-6691-808X">Nicolas Oudart</a>)
**Repository:** https://github.com/NicOudart/PyBWE
**Branch with paper.md** (empty if default branch): main
**Version:** 2024.02
