### 2024-02-04 #6319 @NicOudart
> @NicOudart — Thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

Thank you @dfm, of course I understand. Just to let you know, I said the version to be reviewed was 2024.02, but I detected today a minor mistake in one of my performance test scripts, and I fixed it. I named the new version 2024.02.2. Is it possible to change it? (If not, the rest of the software is unchanged and can be reviewed as is). My apologies.
