### 2024-02-05 #6319 @dfm
@NicOudart — No problem at all. For JOSS reviews we typically review the current default branch of the code unless the authors explicitly ask for us to review a specific version. At the end of the review process, we typically ask for a snapshot to be made of the final version of the code after review, and that's the artifact that gets associated with the paper. Hope this helps!
