### 2024-02-05 5993704
Add JOSS badge to the README.
