### 2024-03-07 cf346da
Modification of the signal model from a sum of negative exponential to a sum of positive exponential, for simplification.
