+++ joss/paper.bib
@@ -1,0 +1,198 @@
+@article{Ciarletti:2017,
+  title = {The WISDOM Radar: Unveiling the Subsurface Beneath the ExoMars Rover and Identifying the Best Locations for Drilling},
+  volume = {17},
+  ISSN = {1557-8070},
+  url = {http://dx.doi.org/10.1089/ast.2016.1532},
+  DOI = {10.1089/ast.2016.1532},
+  number = {6–7},
+  journal = {Astrobiology},
+  publisher = {Mary Ann Liebert Inc},
+  author = {Ciarletti,  Valérie and Clifford,  Stephen and Plettemeier,  Dirk and Le Gall,  Alice and Hervé,  Yann and Dorizon,  Sophie and Quantin-Nataf,  Cathy and Benedix,  Wolf-Stefan and Schwenzer,  Susanne and Pettinelli,  Elena and Heggy,  Essam and Herique,  Alain and Berthelier,  Jean-Jacques and Kofman,  Wlodek and Vago,  Jorge L. and Hamran,  Svein-Erik and the WISDOM Team},
+  year = {2017},
+  month = jul,
+  pages = {565–584}
+}
+ 
+@techreport{Cuomo:1992,
+ title = {A bandwidth extrapolation technique for improved range resolution of coherent radar data},
+ author = {Cuomo, Kevin M.},
+ institution = {Lincoln Laboratory MIT},
+ year  = {1992}
+}
+
+@article{Eide:2022,
+  title = {Radar Attenuation in the Shallow Martian Subsurface: RIMFAX Time Frequency Analysis and Constant Q Characterization Over Jezero Crater Floor},
+  volume = {50},
+  ISSN = {1944-8007},
+  url = {http://dx.doi.org/10.1029/2022GL101429},
+  DOI = {10.1029/2022gl101429},
+  number = {7},
+  journal = {Geophysical Research Letters},
+  publisher = {American Geophysical Union (AGU)},
+  author = {Eide,  Sigurd and Casademont,  Titus M. and Berger,  Tor and Dypvik,  Henning and Shoemaker,  Emileigh S. and Hamran,  Svein Erik},
+  year = {2022},
+  month = dec 
+}
+
+@article{Gambacorta:2022,
+  title = {UWB Processing Applied to Multifrequency Radar Sounders: The Case of MARSIS and Comparison With SHARAD},
+  volume = {60},
+  ISSN = {1558-0644},
+  url = {http://dx.doi.org/10.1109/TGRS.2022.3216893},
+  DOI = {10.1109/tgrs.2022.3216893},
+  journal = {IEEE Transactions on Geoscience and Remote Sensing},
+  publisher = {Institute of Electrical and Electronics Engineers (IEEE)},
+  author = {Gambacorta,  L. and Raguso,  M. C. and Mastrogiuseppe,  M. and Seu,  R.},
+  year = {2022},
+  pages = {1–14}
+}
+
+@Article{Harris:2020,
+ title = {Array programming with {NumPy}},
[…]
