### 2024-04-03 #6319 @letiziagambacorta1
***@***.*** <https://github.com/mbobra> I think all of them are a a good
choices. M. Raguso is not from my institution. She is JPL/Calthec.

On Wed, 3 Apr 2024 at 05:15, Nicolas Oudart ***@***.***>
wrote:

> 👋 @NicOudart <https://github.com/NicOudart> I'm still working on finding
> the second referee -- if you have further suggestions, let me know.
>
> ***@***.*** <https://github.com/mbobra> I have a few suggestions, but no
> certainty that they have a GitHub account: Sigurd Eide from University of
> Oslo, Valerio Poggiali from Cornell University. Also, Maria Carmela Raguso,
> but if I'm correct, she is in the same institution as Letizia.
>
