### 2024-04-20 #6622 @letiziagambacorta1
Dear Monica,
I just still did not manage to start the review. I will try to provide it
to you no later than the end of the next week.

Il giorno sab 20 apr 2024 alle ore 00:04 Monica Bobra <
***@***.***> ha scritto:

> @letiziagambacorta1 <https://github.com/letiziagambacorta1> Did you find
> the review materials okay? Please let me know if you need any help!
>
