### 2024-04-21 #6622 @letiziagambacorta1
@mbobra I am starting the review! I have some questions. I do not manage to generate my checklist by typing @editorialbot generate my checklist as a comment. The .pdf  rewiew has to be done the same way of the code? By opening an issue in the target directory and linking it in a specific comment in this issue?
