### 2024-04-21 #6622 @mbobra
@letiziagambacorta1 Thank you! Hmmm...I don't know what's wrong. Can you try again?

1. Go to this URL: https://github.com/openjournals/joss-reviews/issues/6622
2. Scroll all the way down the page and find the box that says "Add a comment"
3. Type `@editorialbot generate my checklist`
