### 2024-04-22 1c6df51
Adding a tight layout option to all example scripts, to avoid text overlap in figures. Update of the documentation accordingly.
