+++ examples/script_example_PyBWE_BWE.py
@@ -109,4 +109,5 @@ plt.ylabel('Normalized amplitude')
 plt.legend(['Original radar sounding','Radar sounding after BWE'], loc ='best')
 plt.title('BWE application')
 plt.grid()
+plt.tight_layout()
 plt.show()
