+++ docs/PyBWE.md
@@ -317,6 +317,7 @@ plt.ylabel('Normalized amplitude')
 plt.legend(['Original radar sounding','Radar sounding after BWE'], loc ='best')
 plt.title('BWE application')
 plt.grid()
+plt.tight_layout()
 plt.show()
 ~~~
 
@@ -386,6 +387,7 @@ plt.ylabel('Amplitude')
 plt.legend(['Expected spectrum after extrapolation (real part)','Expected spectrum after extrapolation (modulus)'], loc ='best')
 plt.title('Expected spectrum after extrapolation')
 plt.grid()
+plt.tight_layout()
 plt.show()
 ~~~
 
@@ -415,6 +417,7 @@ plt.ylabel('Amplitude')
 plt.legend(['Original radar spectrum (real part)','Original radar spectrum (modulus)','Backward extrapolation (real part)', 'Backward extrapolation (modulus)', 'Forward extrapolation (real part)', 'Forward extrapolation (modulus)'], loc ='best')
 plt.title('Spectrum AR-extrapolation')
 plt.grid()
+plt.tight_layout()
 plt.show()
 ~~~
 
