### 2024-04-22 baf6644
For all example scripts, modification of the figures size to avoid overlapping plots with legends. Update of the documentation and figures accordingly.
