+++ examples/script_example_PyBWE_extrapolation.py
@@ -102,14 +102,16 @@ freq_vect_expected = np.linspace(freq_vect[0]-(df*(Nextra-2)),freq_vect[-1]+(df*
 spec_vect_expected = np.exp(1j*4*pi*dist_target1*freq_vect_expected/3e8)+np.exp(1j*4*pi*dist_target2*freq_vect_expected/3e8)
 
 #Display the expected spectrum after extrapolation:
+plt.figure(figsize=(10,5))
 plt.plot(freq_vect_expected*1e-9,np.real(spec_vect_expected),'r-')
 plt.plot(freq_vect_expected*1e-9,abs(spec_vect_expected),'k-')
 plt.axvspan((freq_vect[0]-(df*(Nextra-2)))*1e-9, (freq_vect[0]-df)*1e-9, facecolor='k', alpha=0.1)
 plt.axvspan((freq_vect[-1]+df)*1e-9, (freq_vect[-1]+(df*Nextra))*1e-9, facecolor='k', alpha=0.1)
 plt.xlim([(freq_vect[0]-(df*(Nextra-2)))*1e-9,(freq_vect[-1]+(df*Nextra))*1e-9])
+plt.ylim([-2.1,2.1])
 plt.xlabel('Frequency (GHz)')
 plt.ylabel('Amplitude')
-plt.legend(['Expected spectrum after extrapolation (real part)','Expected spectrum after extrapolation (modulus)'], loc ='best')
+plt.legend(['Real part','Modulus'], loc ='upper left')
 plt.title('Expected spectrum after extrapolation')
 plt.grid()
 plt.tight_layout()
@@ -120,6 +122,7 @@ freq_vect_forward = np.linspace(freq_vect[-1]+df,freq_vect[-1]+(df*Nextra),Nextr
 freq_vect_backward = np.linspace(freq_vect[0]-(df*(Nextra-2)),freq_vect[0]-df,Nextra)
 
 #Display the original radar spectrum and the extrapolation:
+plt.figure(figsize=(10,5))
 plt.plot(freq_vect*1e-9,np.real(spec_vect),'r-')
 plt.plot(freq_vect*1e-9,abs(spec_vect),'k-')
 plt.plot(freq_vect_backward*1e-9,np.real(spec_vect_backward),'y-')
@@ -129,10 +132,11 @@ plt.plot(freq_vect_forward*1e-9,abs(spec_vect_forward),'b-')
 plt.axvspan((freq_vect[0]-(df*(Nextra-2)))*1e-9, (freq_vect[0]-df)*1e-9, facecolor='k', alpha=0.1)
 plt.axvspan((freq_vect[-1]+df)*1e-9, (freq_vect[-1]+(df*Nextra))*1e-9, facecolor='k', alpha=0.1)
 plt.xlim([(freq_vect[0]-(df*(Nextra-2)))*1e-9,(freq_vect[-1]+(df*Nextra))*1e-9])
+plt.ylim([-3,2.5])
 plt.xlabel('Frequency (GHz)')
 plt.ylabel('Amplitude')
-plt.legend(['Original radar spectrum (real part)','Original radar spectrum (modulus)','Backward extrapolation (real part)', 'Backward extrapolation (modulus)', 'Forward extrapolation (real part)', 'Forward extrapolation (modulus)'], loc ='best')
 plt.title('Spectrum AR-extrapolation')
 plt.grid()
 plt.tight_layout()
+plt.legend(['Original radar spectrum (real part)','Original radar spectrum (modulus)','Backward extrapolation (real part)', 'Backward extrapolation (modulus)', 'Forward extrapolation (real part)', 'Forward extrapolation (modulus)'],loc ='lower center',ncol=3)
 plt.show()
