+++ examples/script_example_PyPBWE_PBWE.py
@@ -123,10 +123,13 @@ output_pbwe, time_pbwe_vect = PyPBWE.PBWE(spec_mat,df,extra_factor,model_order,z
 time_vect = np.linspace(0,1/df,zp_factor*np.shape(spec_mat)[1])
 
 #Display the original radar sounding and the PBWE version for channel 00:
+plt.figure(figsize=(12.75,5))
+
 plt.subplot(1,2,1)
 plt.plot(time_vect*1e9,abs(1.85*np.fft.fft(spec_mat[0,:]*np.hamming(len(spec_mat[0,:])),zp_factor*len(spec_mat[0,:])))/len(spec_mat[0,:]),'k-')
 plt.plot(time_pbwe_vect*1e9,abs(output_pbwe[0,:]),'r-')
 plt.xlim([5,9])
+plt.ylim([-0.05,1.3])
 plt.xlabel('Time delays (ns)')
 plt.ylabel('Normalized amplitude')
 plt.legend(['Original radar sounding','Radar sounding after PBWE'], loc ='best')
@@ -138,6 +141,7 @@ plt.subplot(1,2,2)
 plt.plot(time_vect*1e9,abs(1.85*np.fft.fft(spec_mat[1,:]*np.hamming(len(spec_mat[1,:])),zp_factor*len(spec_mat[1,:])))/len(spec_mat[1,:]),'k-')
 plt.plot(time_pbwe_vect*1e9,abs(output_pbwe[1,:]),'r-')
 plt.xlim([5,9])
+plt.ylim([-0.05,1.3])
 plt.xlabel('Time delays (ns)')
 plt.ylabel('Normalized amplitude')
 plt.legend(['Original radar sounding','Radar sounding after PBWE'], loc ='best')
