+++ examples/script_example_PyPBWE_extrapolation.py
@@ -133,27 +133,31 @@ spec_vect_expected_11 = np.exp(1j*4*pi*dist_target1*freq_vect_expected/3e8)-np.e
 
 #Display the expected spectrum after extrapolation for the 00 and 11
 #channels:
+plt.figure(figsize=(10,6))
+
 plt.subplot(2,1,1)
 plt.plot(freq_vect_expected*1e-9,np.real(spec_vect_expected_00),'r-')
 plt.plot(freq_vect_expected*1e-9,abs(spec_vect_expected_00),'k-')
 plt.axvspan((freq_vect[0]-(df*(Mextra-2)))*1e-9, (freq_vect[0]-df)*1e-9, facecolor='k', alpha=0.1)
 plt.axvspan((freq_vect[-1]+df)*1e-9, (freq_vect[-1]+(df*Mextra))*1e-9, facecolor='k', alpha=0.1)
 plt.xlim([(freq_vect[0]-(df*(Mextra-2)))*1e-9,(freq_vect[-1]+(df*Mextra))*1e-9])
+plt.ylim([-2.5,3])
 plt.xlabel('Frequency (GHz)')
 plt.ylabel('Amplitude')
-plt.legend(['Expected spectrum after extrapolation (real part)','Expected spectrum after extrapolation (modulus)'], loc ='best')
+plt.legend(['Real part','Modulus'],loc='upper left',ncol=2)
 plt.title('Expected spectrum after extrapolation - Polar channel 00')
 plt.grid()
 
 plt.subplot(2,1,2)
 plt.plot(freq_vect_expected*1e-9,np.real(spec_vect_expected_11),'r-')
 plt.plot(freq_vect_expected*1e-9,abs(spec_vect_expected_11),'k-')
-plt.axvspan((freq_vect[0]-(df*(Mextra-2)))*1e-9, (freq_vect[0]-df)*1e-9, facecolor='k', alpha=0.1)
-plt.axvspan((freq_vect[-1]+df)*1e-9, (freq_vect[-1]+(df*Mextra))*1e-9, facecolor='k', alpha=0.1)
+plt.axvspan((freq_vect[0]-(df*(Mextra-2)))*1e-9, (freq_vect[0]-df)*1e-9,facecolor='k',alpha=0.1)
+plt.axvspan((freq_vect[-1]+df)*1e-9, (freq_vect[-1]+(df*Mextra))*1e-9,facecolor='k',alpha=0.1)
 plt.xlim([(freq_vect[0]-(df*(Mextra-2)))*1e-9,(freq_vect[-1]+(df*Mextra))*1e-9])
+plt.ylim([-2.5,3])
 plt.xlabel('Frequency (GHz)')
 plt.ylabel('Amplitude')
-plt.legend(['Expected spectrum after extrapolation (real part)','Expected spectrum after extrapolation (modulus)'], loc ='best')
+plt.legend(['Real part','Modulus'],loc='upper left',ncol=2)
 plt.title('Expected spectrum after extrapolation - Polar channel 11')
 plt.grid()
 
@@ -166,6 +170,7 @@ freq_vect_backward = np.linspace(freq_vect[0]-(df*(Mextra-2)),freq_vect[0]-df,Me
 
 #Display the original radar spectrum and the extrapolation for the 00 and 11
 #channels:
+plt.figure(figsize=(11,6))
 
 plt.subplot(2,1,1)
 plt.plot(freq_vect*1e-9,np.real(spec_mat[0,:]),'r-')
@@ -174,9 +179,11 @@ plt.plot(freq_vect_backward*1e-9,np.real(spec_mat_extra_backward[0,:]),'y-')
[…]

[…]
