### 2024-04-22 215b1f0
Fixing a typo in the PySSBWE documentation.
