### 2024-04-23 960ab9e
Add a link to the PyBWE website at the beginning of the README.
