### 2024-04-23 #6622 @NicOudart
For your information, following an issue raised by [kathlandgren](https://github.com/kathlandgren) concerning the figures outputs of PyBWE example scripts, I updated the article proof with the new figure format.
