### 2024-05-03 92334d9
Update of the JOSS paper and references: - Replacement of "better" by "enhanced". - Clarification of the distinction between the type of model and the method to determine the model's coefficients. - Adding the use of the BWI for EMI removal in SHARAD data. - Adding the latest reference on BWE applied to SHARAD (Raguso et al., 2023) and the (Piazzo et al., 2019) reference for EMI removal by BWI.
