+++ docs/joss/paper.bib
@@ -140,6 +140,20 @@
  url = {https://doi.org/10.5281/zenodo.3509134}
 }
 
+@article{Piazzo:2019,
+  title = {Signal enhancement for planetary radar sounders},
+  volume = {55},
+  ISSN = {1350-911X},
+  url = {http://dx.doi.org/10.1049/el.2018.7284},
+  DOI = {10.1049/el.2018.7284},
+  number = {3},
+  journal = {Electronics Letters},
+  publisher = {Institution of Engineering and Technology (IET)},
+  author = {Piazzo,  L. and Raguso,  M.C. and Seu,  R. and Mastrogiuseppe,  M.},
+  year = {2019},
+  month = feb,
+  pages = {153–155}
+}
  
 @techreport{Piou:1999,
  title = {A State-Space Technique for Ultrawide-Bandwidth Coherent Processing},
@@ -157,6 +171,19 @@
   author = {Raguso,  Maria Carmela and Mastrogiuseppe,  Marco and Seu,  Roberto and Piazzo,  Lorenzo},
   year = {2018},
   month = jun 
+}
+
+@article{Raguso:2023,
+  title = {Range resolution enhancement of SHAllow RADar (SHARAD) data via bandwidth extrapolation technique: Enabling new features detection and improving geophysical investigation},
+  ISSN = {0019-1035},
+  url = {http://dx.doi.org/10.1016/j.icarus.2023.115803},
+  DOI = {10.1016/j.icarus.2023.115803},
+  journal = {Icarus},
+  publisher = {Elsevier BV},
+  author = {Raguso,  M.C. and Mastrogiuseppe,  M. and Gambacorta,  L. and Di Achille,  G. and Seu,  R.},
+  year = {2023},
+  month = sep,
+  pages = {115803}
 }
 
 @inproceedings{Suwa:2003,
