+++ docs/joss/paper.bib
@@ -7,7 +7,7 @@
   number = {6–7},
   journal = {Astrobiology},
   publisher = {Mary Ann Liebert Inc},
-  author = {Ciarletti,  Valérie and Clifford,  Stephen and Plettemeier,  Dirk and Le Gall,  Alice and Hervé,  Yann and Dorizon,  Sophie and Quantin-Nataf,  Cathy and Benedix,  Wolf-Stefan and Schwenzer,  Susanne and Pettinelli,  Elena and Heggy,  Essam and Herique,  Alain and Berthelier,  Jean-Jacques and Kofman,  Wlodek and Vago,  Jorge L. and Hamran,  Svein-Erik and the WISDOM Team},
+  author = {Ciarletti,  V. and Clifford,  S. and Plettemeier,  D. and Le Gall,  A. and Hervé,  Y. and Dorizon,  S. and Quantin-Nataf,  C. and Benedix,  W.-S. and Schwenzer,  S. and Pettinelli,  E. and Heggy,  E. and Herique,  A. and Berthelier,  J.J. and Kofman,  W. and Vago,  J.L. and Hamran,  S.Erik. and the WISDOM Team},
   year = {2017},
   month = jul,
   pages = {565–584}
@@ -15,7 +15,7 @@
  
 @techreport{Cuomo:1992,
  title = {A bandwidth extrapolation technique for improved range resolution of coherent radar data},
- author = {Cuomo, Kevin M.},
+ author = {Cuomo, K.M.},
  institution = {Lincoln Laboratory MIT},
  year  = {1992}
 }
@@ -29,7 +29,7 @@
   number = {7},
   journal = {Geophysical Research Letters},
   publisher = {American Geophysical Union (AGU)},
-  author = {Eide,  Sigurd and Casademont,  Titus M. and Berger,  Tor and Dypvik,  Henning and Shoemaker,  Emileigh S. and Hamran,  Svein Erik},
+  author = {Eide,  S. and Casademont,  T.M. and Berger,  T. and Dypvik,  H. and Shoemaker,  E.S. and Hamran,  S.E.},
   year = {2022},
   month = dec 
 }
@@ -42,7 +42,7 @@
   DOI = {10.1109/tgrs.2022.3216893},
   journal = {IEEE Transactions on Geoscience and Remote Sensing},
   publisher = {Institute of Electrical and Electronics Engineers (IEEE)},
-  author = {Gambacorta,  L. and Raguso,  M. C. and Mastrogiuseppe,  M. and Seu,  R.},
+  author = {Gambacorta,  L. and Raguso,  M.C. and Mastrogiuseppe,  M. and Seu,  R.},
   year = {2022},
   pages = {1–14}
 }
@@ -62,7 +62,7 @@
 }
 
 @Article{Hunter:2007,
-  Author = {Hunter, J. D.},
+  Author = {Hunter, J.D.},
   Title = {Matplotlib: A 2D graphics environment},
   Journal = {Computing in Science \& Engineering},
   Volume = {9},
@@ -97,7 +97,7 @@
   number = {5},
   journal = {Geophysical Research Letters},
   publisher = {American Geophysical Union (AGU)},
-  author = {Mastrogiuseppe,  Marco and Poggiali,  Valerio and Hayes,  Alexander and Lorenz,  Ralph and Lunine,  Jonathan and Picardi,  Giovanni and Seu,  Roberto and Flamini,  Enrico and Mitri,  Giuseppe and Notarnicola,  Claudia and Paillou,  Philippe and Zebker,  Howard},
[…]
