### 2024-05-03 a819c5a
Rephrase a sentence in the JOSS paper.
