### 2024-05-03 72d491b
Update of the PySSBWE documentation to make clear that the number of selected singular values corresponds to the order of the model, ideally equal to the number target echoes in the radar signal.
