### 2024-05-03 #6622 @NicOudart
For your information, following an issue raised by [letiziagambacorta1](https://github.com/letiziagambacorta1) concerning her review of the article, I updated the article proof with minor corrections.
