### 2024-05-22 bad78cb
Minor fix of the figure caption in the JOSS article.
