### 2024-05-22 #6622 @NicOudart
> Almost there, @NicOudart ! Congratulations on an a really nice package and paper. If you can fix the missing DOIs in the paper, please do so. Otherwise (if there is no DOI associated with the reference), that's fine.

Thank you @mbobra ! I made a last minute modification to the paper as I noticed a mistake in the figure's caption (a and b were switched). Concerning the 3 missing DOIs, these references are technical reports from the Lincoln Laboratory (MIT), and do not have a DOI.
