### 2024-05-22 #6622 @mbobra
@NicOudart Okay great! Since we made some modifications, could you please cut a new release of PyBWE, deposit it in Zenodo, and put the link to the Zenodo deposit in this thread? Please make sure that the author list and title on the Zenodo deposit matches the author list and title on the JOSS paper. Then I can move forward with accepting the paper.
