### 2024-05-23 #6622 @NicOudart
> @NicOudart Okay great! Since we made some modifications, could you please cut a new release of PyBWE, deposit it in Zenodo, and put the link to the Zenodo deposit in this thread? Please make sure that the author list and title on the Zenodo deposit matches the author list and title on the JOSS paper. Then I can move forward with accepting the paper.

I cut a new release "2024.05" of PyBWE, with the modified article. I deposited this release to Zenodo, here is the URL: https://zenodo.org/doi/10.5281/zenodo.10259072
However, I am not sure to understand what you mean by "making sure the author list and title on the Zenodo deposit matches the author list and title on the JOSS paper". Do you mean that the paper.md within the release I deposited in Zenodo should have the same title and authors as the draft paper? If so, the title and list of author remained unchanged.
