### 2024-05-24 #6622 @mbobra
> Do you mean that the paper.md within the release I deposited in Zenodo should have the same title and authors as the draft paper?

Yes exactly -- thank you, this is perfect.
