### 2024-05-24 #6622 @mbobra
👋 @openjournals/aass-eics Please note @NicOudart's comment above:

> Concerning the 3 missing DOIs, these references are technical reports from the Lincoln Laboratory (MIT), and do not have a DOI.
